该版本的多分类版本已经可以跑通！多分类的损失函数暂时使用softmax

+++ ppocr/data/imaug/make_shrink_map.py
@@ -98,7 +98,7 @@ class MakeShrinkMap(object):
                     
                     
                     
-        data["shrink_class"] = text_class
+        data["shrink_class"] = text_class.astype(np.int64)
         data["shrink_map"] = gt
         data["shrink_mask"] = mask
         return data
